@@ -3,14 +3,14 @@
   "devDependencies": {
     "babel-eslint": "^9.0.0",
     "cross-env": "^5.2.0",
-    "eslint": "^5.6.0",
+    "eslint": "^5.7.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^3.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-prettier": "^2.6.2",
     "eslint-plugin-react": "^7.11.1",
-    "lerna": "^3.2.1",
+    "lerna": "^3.4.3",
     "prettier": "^1.14.3",
     "prettier-eslint": "^8.8.2"
   },
@@ -24,14 +24,19 @@
     ]
   },
   "scripts": {
-    "start": "lerna run --parallel start:dev"
+    "start": "lerna run --parallel start:dev",
+    "build": "lerna run --parallel build",
+    "test": "lerna run --parallel test"
   },
   "name": "diff",
   "dependencies": {
+    "node-pre-gyp": "0.6.x",
+    "firebase": "^5.4.2",
+    "firebase-admin": "^6.0.0",
     "gatsby-cli": "^1.1.58"
   },
   "engines": {
-    "npm": ">5.0.0",
+    "npm": ">6.4.0",
     "node": ">8.0.0"
   }
 }
