@@ -1,25 +1,9 @@
 import { combineEpics, ofType } from 'redux-observable';
 import { of } from 'rxjs';
-import { catchError, switchMap, map, flatMap } from 'rxjs/operators';
+import { catchError, switchMap, flatMap } from 'rxjs/operators';
 import types from './types';
 import actions from './actions';
 import { actions as userActions } from '../users';
-import { types as userTypes } from '../session';
-
-const getInvitesEpic = (action$, state$, { api }) =>
-  action$.pipe(
-    ofType(userTypes.SELECT_WORKSPACE),
-    switchMap(action =>
-      api.invites.invitesForWorkspace$(action.payload.workspaceId).pipe(
-        map(response => {
-          if (response.type === 'added' || response.type === 'modified') {
-            return actions.addInviteUser(response.data.email);
-          }
-        }),
-        catchError(err => of(actions.addInviteUserFailed(err.message)))
-      )
-    )
-  );
 
 const getWorkspaceByIdEpic = (action$, state$, { api }) =>
   action$.pipe(
@@ -47,4 +31,4 @@ const getWorkspaceByIdEpic = (action$, state$, { api }) =>
     )
   );
 
-export default combineEpics(getWorkspaceByIdEpic, getInvitesEpic);
+export default combineEpics(getWorkspaceByIdEpic);
