@@ -3,5 +3,5 @@ export { default as types } from './types';
 export { default as sources } from './sources';
 export { default as initApi } from './api';
 export { default as browser } from './browser';
-export * from './token';
+export { default as remoteSettings } from './remoteSettings';
 export { default as AuthProvider } from './auth0';
