@@ -17,6 +17,40 @@ const checkSession = async () => {
     : { id_token: null, access_token: null };
 };
 
+/**
+ * Attempts to get a user if one is present in the cookies,
+ * otherwise this will just fail out with an erorr
+ */
+const getUser = async () => {
+  const response = await fetch(`${process.env.API_SERVER}/auth/profile`, {
+    credentials: 'include',
+    mode: 'cors'
+  });
+
+  if (!response.ok) {
+    throw new Error('No logged in user');
+  }
+
+  return response.json();
+};
+
+const logoutUser = async () => {
+  console.warn('user wants to logout');
+};
+
+const getFirebaseToken = async () => {
+  const response = await fetch(`${process.env.API_SERVER}/auth/firebase`, {
+    credentials: 'include',
+    mode: 'cors'
+  });
+
+  if (!response.ok) {
+    throw new Error('No logged in user');
+  }
+
+  return response.json();
+};
+
 const authorize = async (webAuthInstance, state, nonce, redirectUri) => {
   // set browser auth0 authorize from our custom stsate,
   // not sure that this is even needed
@@ -48,6 +82,9 @@ const renewSession = async () => {
 
 export default {
   renewSession,
+  logoutUser,
   authorize,
-  checkSession
+  checkSession,
+  getUser,
+  getFirebaseToken
 };
