@@ -48,9 +48,9 @@ export default class Login extends React.Component {
       )
       .then(() =>
         this.navigateTo(
-          `${process.env.WEB_HOME}/app/login?return_to=${encodeURIComponent(
+          `${process.env.WEB_HOME}/app/login?origin=${encodeURIComponent(
             `chrome-extension://${browser.runtime.id}/html/login-return.html`
-          )}&state=${state}&nonce=${nonce}`
+          )}`
         )
       );
   };
