@@ -1,6 +1,5 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import jwtDecode from 'jwt-decode';
+import browser from '@diff/common/dist/browser';
 import { StyleBoundary } from '@diff/shared-components';
 
 import styled from 'styled-components';
@@ -14,63 +13,36 @@ const Container = styled.div`
 `;
 
 export default class Popup extends React.Component {
-  static propTypes = {
-    authProvider: PropTypes.object.isRequired
-  };
-
   state = {
-    accessToken: null
+    user: null
   };
 
-  async componentDidMount() {
-    const {
-      props: { authProvider }
-    } = this;
-
-    try {
-      const {
-        access_token: accessToken
-      } = await authProvider.checkAndRenewSession();
-      this.setState({ accessToken });
-    } catch (error) {
-      console.error('No Session available');
-    }
+  componentDidMount() {
+    // attempt to get our user, if no user
+    // object then we probably encountered a no login
+    // state
+    browser.auth.getUser().then(user => {
+      this.setState({
+        user
+      });
+    });
   }
 
-  logout = () => {
-    const {
-      props: { authProvider }
-    } = this;
-
-    authProvider.logout();
-
-    // Remove the idToken from storage
-    localStorage.clear();
-    this.setState({ accessToken: null });
-    chrome.browserAction.setIcon({ path: '../images/inactive_icon_128.png' });
+  logout = async () => {
+    await browser.auth.logoutUser();
   };
-
-  // The user is logged in if their token isn't expired
-  isLoggedIn = token => jwtDecode(token).exp > Date.now() / 1000;
 
   render() {
     const {
-      props: { authProvider },
-      state: { accessToken },
+      state: { user },
       logout
     } = this;
 
     return (
       <StyleBoundary>
         <Container>
-          {accessToken && (
-            <User
-              accessToken={accessToken}
-              onLogout={logout}
-              authProvider={authProvider}
-            />
-          )}
-          {!accessToken && <Login />}
+          {user && <User user={user} onLogout={logout} />}
+          {!user && <Login />}
         </Container>
       </StyleBoundary>
     );
