@@ -3,6 +3,8 @@ import { Switch, Route } from 'react-router';
 
 import CreateWorkspace from './CreateWorkspace';
 import ListWorkspaces from './ListWorkspaces';
+import ListCollaborators from './ListCollaborators';
+import AddCollaborator from './AddCollaborator';
 
 const Placeholder = ({ text }) => (
   <h1>
@@ -18,11 +20,11 @@ const Annotations = () => (
     <Switch>
       <Route exact path="/workspace/create" component={CreateWorkspace} />
       <Route exact path="/workspace/find" component={Finder} />
-      <Route exact path="/workspace/:id" component={Placeholder} />
+      <Route exact path="/workspace/:id" component={ListCollaborators} />
       <Route
         exact
         path="/workspace/:id/addCollaborator"
-        component={Placeholder}
+        component={AddCollaborator}
       />
 
       <Route exact path="/workspace" component={ListWorkspaces} />
