@@ -6,13 +6,14 @@ import Row from '../../components/Row';
 
 const Content = ({ history }) => (
   <div>
-    <Row onClick={() => history.push('/workspace/find')}>Find a workspace</Row>
     <Row onClick={() => history.push('/workspace/create')}>Add a workspace</Row>
     <Workspaces>
       {workspaces => (
         <React.Fragment>
           {workspaces.map(workspace => (
-            <Row>{workspace.name}</Row>
+            <Row onClick={() => history.push(`/workspace/${workspace.id}`)}>
+              {workspace.name}
+            </Row>
           ))}
         </React.Fragment>
       )}
