@@ -1,19 +1,20 @@
-const ADD_WORKSPACE_USER_REQUEST =
-  "@diff/widgets/workspaces/addWorkspaceUserRequest";
-const ADD_WORKSPACE_USER_SUCCESS =
-  "@diff/widgets/workspaces/addWorkspaceUserSuccess";
-const ADD_WORKSPACE_USER_FAILED =
-  "@diff/widgets/workspaces/addWorkspaceUserFailed";
-const CREATE_WORKSPACE = "@diff/widgets/workspaces/createWorkspace";
+const INVITE_TO_WORKSPACE_REQUEST =
+  '@diff/widgets/workspaces/addWorkspaceUserRequest';
+const INVITE_TO_WORKSPACE_SUCCESS =
+  '@diff/widgets/workspaces/addWorkspaceUserSuccess';
+const INVITE_TO_WORKSPACE_FAILED =
+  '@diff/widgets/workspaces/addWorkspaceUserFailed';
+
+const CREATE_WORKSPACE = '@diff/widgets/workspaces/createWorkspace';
 const CREATE_WORKSPACE_SUCCESS =
-  "@diff/widgets/workspaces/createWorkspaceSuccess";
+  '@diff/widgets/workspaces/createWorkspaceSuccess';
 const CREATE_WORKSPACE_FAILED =
-  "@diff/widgets/workspaces/createWorkspaceFailed";
+  '@diff/widgets/workspaces/createWorkspaceFailed';
 
 export default {
-  ADD_WORKSPACE_USER_REQUEST,
-  ADD_WORKSPACE_USER_SUCCESS,
-  ADD_WORKSPACE_USER_FAILED,
+  INVITE_TO_WORKSPACE_REQUEST,
+  INVITE_TO_WORKSPACE_SUCCESS,
+  INVITE_TO_WORKSPACE_FAILED,
   CREATE_WORKSPACE,
   CREATE_WORKSPACE_SUCCESS,
   CREATE_WORKSPACE_FAILED
