@@ -1,10 +1,11 @@
 import jwt from 'express-jwt';
 import jwks from 'jwks-rsa';
 
-import { admin, db } from '../../firestore';
+import { client as firestoreClient } from '../../firestore';
 import logging from '../../logging';
-import * as userCollection from '../../firestore/users';
+import * as userCollection from '../../firestore/models/users';
 
+const { admin, db } = firestoreClient;
 const request = require('request');
 
 const config = {
@@ -24,6 +25,19 @@ const jwtCheck = jwt({
   issuer: `https://${config.AUTH0_DOMAIN}/`,
   algorithm: 'RS256'
 });
+
+const cookieToken = (req, res, next) => {
+  console.log('cookies', req.cookies._df_id_token);
+  const token = req.cookies._df_id_token;
+  if (!token) {
+    return res.status(401).send('sorry not authorized');
+  }
+
+  req.headers.authorization = `Bearer ${token}`;
+
+  console.log('headers now', req.headers.authorization);
+  next();
+};
 
 const userProfileSync = async (req, res, next) => {
   logging.info('------------[first-time check] ----------');
@@ -47,6 +61,7 @@ const userProfileSync = async (req, res, next) => {
 
 // GET object containing Firebase custom token
 export const login = [
+  cookieToken,
   jwtCheck,
   userProfileSync,
   (req, res) => {
@@ -104,10 +119,20 @@ export const refresh = (req, res) => {
   });
 };
 
+export const profile = [
+  cookieToken,
+  jwtCheck,
+  (req, res) => {
+    res.status(200).json(req.user);
+  }
+];
+
 export const codeGrantAuthorize = [
   (req, res, next) => {
+    console.log('received a login event', req.query);
+
     const {
-      query: { code, redirectUri }
+      query: { code }
     } = req;
 
     const options = {
@@ -119,15 +144,17 @@ export const codeGrantAuthorize = [
         client_id: process.env.AUTH0_CLIENTID,
         client_secret: process.env.AUTH0_CLIENT_SECRET,
         code,
-        redirect_uri: redirectUri
+        redirect_uri: process.env.API_SERVER
       },
       json: true
     };
 
     request(options, (error, response, body) => {
       if (error) {
+        console.error('error', error);
         next(error);
       }
+      console.log(body);
 
       req.headers.authorization = `Bearer ${body.id_token}`;
       req._user = body;
@@ -137,7 +164,14 @@ export const codeGrantAuthorize = [
   jwtCheck,
   userProfileSync,
   (req, res) => {
-    res.status(200).send(req._user);
+    const body = req._user;
+
+    res.cookie('_df_id_token', body.id_token, { httpOnly: true });
+    res.cookie('_df_access_token', body.access_token, { httpOnly: true });
+    res.cookie('_df_refresh_token', body.refresh_token, { httpOnly: true });
+    res.cookie('_df_expires_in', body.refresh_token, { httpOnly: true });
+    res.cookie('_df_token_type', body.token_type, { httpOnly: true });
+    res.redirect(301, req.query.origin);
   }
 ];
 
