@@ -3,21 +3,37 @@
  *
  * See: https://www.gatsbyjs.org/docs/node-apis/
  */
+const Dotenv = require('dotenv-webpack');
 
-require('dotenv').config({
-  path: `.env.${process.env.NODE_ENV}`,
-})
+const config = `../../.env.${process.env.NODE_ENV}`;
+require('dotenv').config({ path: config });
 
 // You can delete this file if you're not using it
 exports.onCreatePage = async ({ page, actions }) => {
-  const { createPage } = actions
+  const { createPage } = actions;
 
   // page.matchPath is a special key that's used for matching pages
   // only on the client.
   if (page.path.match(/^\/app/)) {
-    page.matchPath = `/app/*`
+    page.matchPath = `/app/*`;
 
     // Update the page.
-    createPage(page)
+    createPage(page);
   }
-}
+};
+
+exports.onCreateWebpackConfig = ({
+  stage,
+  getConfig,
+  rules,
+  loaders,
+  actions
+}) => {
+  actions.setWebpackConfig({
+    plugins: [
+      new Dotenv({
+        path: config
+      })
+    ]
+  });
+};
