@@ -1,4 +1,4 @@
-import * as firebase from 'firebase/app';
+import { initApi } from '@diff/common';
 import 'firebase/firestore';
 import 'firebase/auth';
 
@@ -14,25 +14,8 @@ export default class FirestoreApi {
   }
 
   static connect() {
-    console.log('[plugin - firebase] initializing connection');
-
-    // connect to firebase
-    const config = {
-      apiKey: process.env.FIREBASE_API_KEY,
-      authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-      databaseURL: process.env.FIREBASE_DATABASE_URL,
-      projectId: process.env.FIREBASE_PROJECT_ID,
-      storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-      messagingSenderId: process.env.FIREBASE_SENDER_ID
-    };
-
-    firebase.initializeApp(config);
-
-    const firestore = firebase.firestore();
-    const settings = { timestampsInSnapshots: true };
-    firestore.settings(settings);
-
-    return new FirestoreApi(firestore);
+    const firestoreApi = initApi();
+    return new FirestoreApi(firestoreApi);
   }
 
   get currentPage() {
@@ -147,8 +130,6 @@ export default class FirestoreApi {
   }
 
   async updateUser(key, value) {
-    /* eslint-disable */
-    debugger;
     return this.db
       .collection('users')
       .doc(this.currentUid)
