@@ -10,8 +10,19 @@ import AttachFeature from './features/Attach';
 import TrackFeature from './features/Track';
 
 import DiffHeader from './diff.gif';
-import DiffLogo from './Diff.svg';
-import DiffDarkLogo from './Diff2.svg';
+import DiffLogo from './diff.svg';
+import DiffDarkLogo from './diff2.svg';
+
+if (typeof window !== 'undefined') {
+  if (window.navigator && navigator.serviceWorker) {
+    navigator.serviceWorker.getRegistrations().then(registrations => {
+      for (const registration of registrations) {
+        registration.unregister();
+      }
+    });
+    console.log('Cache cleared');
+  }
+}
 
 const Container = styled.div`
   color: #130c3b;
